# Conflicts:
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/css/resetpasswordpage.css
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/css/setpasswordpage.css
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/css/signinpage.css
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/contentpage.js
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/resetpasswordpage.js
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/setpasswordpage.js
#	be/src/Project/Services/Identity/Identity.Api/wwwroot/dist/js/signinpage.js
#	fe/src/project/Account/areas/Accounts/pages/SignIn/SignInPage.js

+++ be/src/Infrastructure/PageContent/Foundation.PageContent/Components/CarouselGrids/ViewModels/CarouselGridCarouselItemViewModel.cs
@@ -15,8 +15,8 @@ namespace Foundation.PageContent.Components.CarouselGrids.ViewModels
         public string Subtitle { get; set; }
         public string Ean { get; set; }
         public string OutletTitle { get; set; }
-        public int? AvailableQuantity { get; set; }
-        public int? AvailableOutletQuantity { get; set; }
+        public double? AvailableQuantity { get; set; }
+        public double? AvailableOutletQuantity { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
         public DateTime? CreatedDate { get; set; }
         public IEnumerable<SourceViewModel> Sources { get; set; }
