+++ be/src/Project/Services/Inventory/Inventory.Api/IntegrationEventsModels/BasketCheckoutProductEventModel.cs
@@ -5,6 +5,6 @@ namespace Inventory.Api.IntegrationEventsModels
     public class BasketCheckoutProductEventModel
     {
         public Guid? ProductId { get; set; }
-        public int BookedQuantity { get; set; }
+        public double BookedQuantity { get; set; }
     }
 }
