+++ be/src/Project/Services/Inventory/Inventory.Api/v1/ResponseModels/InventorySumResponseModel.cs
@@ -9,8 +9,8 @@ namespace Inventory.Api.v1.ResponseModels
         public string ProductName { get; set; }
         public string ProductSku { get; set; }
         public string Ean { get; set; }
-        public int? Quantity { get; set; }
-        public int? AvailableQuantity { get; set; }
+        public double? Quantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public IEnumerable<InventoryDetailsResponseModel> Details { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
